TELEFONIA_API_QM/requirements.txt
python-dotenv
requests
psycopg2-binary
pandas
pytest